Fix contract size by removing unneded code
DELAY_EXPIRATION_TIMES wasn't needed since parameter `delayTime` can
account for the possible delay of the oracle.

The tolerance should consider the margin of a few seconds in the delay
of the oracle.

The expiration of the Policy is set as `expectedArrival + tolerance +
delayTime`.

The scheduled task is set at `expectedArrival + tolerance`.